Se organiza la carpeta de componentes
