+++ WEB/src/components/NavBar/RegistrateNavBar.jsx
