+++ WEB/src/pages/Courses.jsx
@@ -1,13 +1,10 @@
-import React from 'react'
-import  SeacrhCourses  from '../components/SeacrhCourses'
+import React from "react";
+import SeacrhCourses from "../components/SearchCourses/SeacrhCourses";
 
 export const Courses = () => {
   return (
     <div>
-        
-        <SeacrhCourses />
-
-
+      <SeacrhCourses />
     </div>
-  )
-}
+  );
+};
