+++ WEB/src/pages/Home.jsx
@@ -1,9 +1,9 @@
 import "../App.css";
 
-import { Testimonial } from "../components/Testimonial";
+import { Testimonial } from "../components/Testimonials/Testimonial";
 import MenuTutorias from "../components/MenuTutorias";
 import ProfesorDestacado from "../components/ProfesorDestacado";
-import { Banner } from "../components/Banner";
+import { Banner } from "../components/Banner/Banner";
 
 export default function Home() {
   const profesorDatos = {
